refactor: huge chat layout refactor
