+++ swapstreet/app/chat/components/EmptyState.tsx
@@ -1,0 +1,11 @@
+import styles from "../ChatLayout.module.css";
+
+export default function EmptyState() {
+  return (
+    <div className={styles.emptyState}>
+      <p className={styles.emptyStateText}>
+        Select a conversation to start chatting
+      </p>
+    </div>
+  );
+}
